Rewrite EditPersonDescriptorBuilder to EditSourceDescriptorBuilder
