--- src/test/java/seedu/address/testutil/EditPersonDescriptorBuilder.java
@@ -1,88 +1,0 @@
-package seedu.address.testutil;
-
-import java.util.Set;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
-
-import seedu.address.logic.commands.EditCommand;
-import seedu.address.logic.commands.EditCommand.EditSourceDescriptor;
-import seedu.address.model.person.Address;
-import seedu.address.model.person.Email;
-import seedu.address.model.person.Name;
-import seedu.address.model.person.Person;
-import seedu.address.model.person.Phone;
-import seedu.address.model.tag.Tag;
-
-/**
- * A utility class to help with building EditSourceDescriptor objects.
- */
-public class EditPersonDescriptorBuilder {
-
-    private EditCommand.EditSourceDescriptor descriptor;
-
-    public EditPersonDescriptorBuilder() {
-        descriptor = new EditSourceDescriptor();
-    }
-
-    public EditPersonDescriptorBuilder(EditSourceDescriptor descriptor) {
-        this.descriptor = new EditCommand.EditSourceDescriptor(descriptor);
-    }
-
-    /**
-     * Returns an {@code EditSourceDescriptor} with fields containing {@code person}'s details
-     */
-    public EditPersonDescriptorBuilder(Person person) {
-        descriptor = new EditCommand.EditSourceDescriptor();
-        descriptor.setTitle(person.getName());
-        descriptor.setPhone(person.getPhone());
-        descriptor.setEmail(person.getEmail());
-        descriptor.setAddress(person.getAddress());
-        descriptor.setTags(person.getTags());
-    }
-
-    /**
-     * Sets the {@code Name} of the {@code EditSourceDescriptor} that we are building.
-     */
-    public EditPersonDescriptorBuilder withName(String name) {
-        descriptor.setTitle(new Name(name));
-        return this;
-    }
-
-    /**
-     * Sets the {@code Phone} of the {@code EditSourceDescriptor} that we are building.
-     */
-    public EditPersonDescriptorBuilder withPhone(String phone) {
-        descriptor.setPhone(new Phone(phone));
-        return this;
-    }
-
-    /**
-     * Sets the {@code Email} of the {@code EditSourceDescriptor} that we are building.
-     */
-    public EditPersonDescriptorBuilder withEmail(String email) {
-        descriptor.setEmail(new Email(email));
-        return this;
-    }
-
-    /**
-     * Sets the {@code Address} of the {@code EditSourceDescriptor} that we are building.
-     */
-    public EditPersonDescriptorBuilder withAddress(String address) {
-        descriptor.setAddress(new Address(address));
-        return this;
-    }
-
-    /**
-     * Parses the {@code tags} into a {@code Set<Tag>} and set it to the {@code EditSourceDescriptor}
-     * that we are building.
-     */
-    public EditPersonDescriptorBuilder withTags(String... tags) {
-        Set<Tag> tagSet = Stream.of(tags).map(Tag::new).collect(Collectors.toSet());
-        descriptor.setTags(tagSet);
-        return this;
-    }
-
-    public EditSourceDescriptor build() {
-        return descriptor;
-    }
-}
